MVP merge (#1)
* mvp done: implemented sync, added configuration, improved logging, fixed duplication issue

* mvp done: use token instead of user-id, endpoint to get usage stats by app

* fix wrong user id on logs saving

* add missed space

* add cors config (i hate them)

* mvp done: add login, chart with app usage

* mvp done: add docker compose, registration page, improve usability of the webapp

* get rid of spring webflux

* eslint and fix backend issues with create/read and cors

+++ wayd-backend/src/main/java/blck/wayd/model/dto/TrackLogDto.java
@@ -3,5 +3,5 @@ package blck.wayd.model.dto;
 /**
  * Track Log.
  */
-public record TrackLogDto(String appName, int timestamp) {
+public record TrackLogDto(String appName, long timestamp) {
 }
